requirements.txt
Django==4.2.7
mysqlclient==2.2.0
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3
